+ source generator instead of reflection (allows trimming) * clear naming for Upscale/Downscale/Rescale, Upsample/Downsample/Resample, Regenerators + exposed kernel functions for chart drawing, etc. + perf-opt by using a safe-window (are where no bounds checks are needed)

+++ System.Drawing.Extensions/Drawing/ColorDomain/ColorQuantizerAdapter.cs
@@ -64,10 +64,10 @@ public sealed class ColorQuantizerAdapter : IColorQuantizer {
   /// </summary>
   public bool AllowFillingColors { get; }
 
-  public Color[] ReduceColorsTo(byte numberOfColors, IEnumerable<Color> usedColors)
+  public Color[] ReduceColorsTo(ushort numberOfColors, IEnumerable<Color> usedColors)
     => this.ReduceColorsTo(numberOfColors, usedColors.Select(c => (c, 1u)));
 
-  public Color[] ReduceColorsTo(byte numberOfColors, IEnumerable<(Color color, uint count)> histogram) {
+  public Color[] ReduceColorsTo(ushort numberOfColors, IEnumerable<(Color color, uint count)> histogram) {
     if (numberOfColors == 0)
       return [];
